+++ basis/universal_1000_1e-12.dlr
@@ -1,0 +1,61 @@
+# index                       freq                        tau                    fermi n                bose n
+    1               -999.46473080965086            4.1817905496025799e-06              -756              -586
+    2                 -974.218185383172            0.00020142042669396847              -372              -375
+    3               -937.95995186974437            0.00048468787601762251              -291              -272
+    4               -888.89255825490056             0.0015720892148122496              -235              -211
+    5               -798.77258050403202             0.0025809364556702375              -165              -168
+    6               -669.64013367420966             0.0035770890865284089              -126              -136
+    7               -585.16354622498284             0.0049955299019160135               -99              -112
+    8               -513.26746762622361             0.0062404107851877508               -80               -92
+    9               -399.38629025201601              0.008470821826943184               -67               -78
+   10               -319.71391372262485              0.011463269026444754               -57               -59
+   11               -281.02002406512781              0.015222159146612063               -49               -51
+   12               -239.46685076890907               0.02092625417731905               -42               -44
+   13               -207.58996658144758              0.023948461947110495               -35               -35
+   14               -175.30685487399199              0.028588639180469288               -31               -30
+   15               -152.77686043627483              0.038194215109068716               -27               -26
+   16               -131.44545365420697                 0.043826713718498               -23               -22
+   17               -99.846572563004003                0.0518974916453619               -20               -19
+   18               -83.705016709276208              0.065722726827103492               -17               -16
+   19               -70.255006016281953              0.073145443278122849               -14               -13
+   20               -58.622496991859023                0.0837050167092762               -12               -11
+   21               -49.923286281502001                0.1075715215693438               -10               -10
+   22               -36.572721639061427               0.13553314923109094                -9                -9
+   23               -31.550229993699524               0.15277686043627486                -8                -8
+   24               -26.892880392335947                 0.175306854873992                -7                -7
+   25               -19.097107554534354               0.22222313956372514                -6                -6
+   26               -13.888946222732821               0.28102002406512783                -5                -5
+   27               -9.9910598038320266               0.31971391372262487                -4                -4
+   28               -6.9444731113664107                 0.350613709747984                -3                -3
+   29               -2.9935577433888119               0.43028608627737519                -2                -2
+   30              0.051826045414935962               0.46897997593487217                -1                -1
+   31                1.5191115556832055               0.51289090730841402                 0                 0
+   32                3.2435932278351185               0.58482006683710486                 1                 1
+   33                7.8875574984248811                 0.649386290252016                 2                 2
+   34                11.463269026444753               0.69444627912745027                 3                 3
+   35                15.775114996849762               0.73710909269158598                 4                 4
+   36                21.913356859248999               0.77129088655624567                 5                 5
+   37                29.311248495929512               0.80841230442311607                 6                 6
+   38                33.883287307772733                0.8401430431386876                 7                 7
+   39                39.964239215328107               0.86446685076890906                 8                 8
+   40                51.897491645361896               0.88275500601628198                 9                 9
+   41                67.766574615545466                 0.900153427436996                11                11
+   42                79.928478430656213               0.91629498329072379                13                13
+   43                 95.79384778844198               0.93584156654672201                15                15
+   44                 128.3168669065559               0.94444421510906873                18                18
+   45                146.29088655624571               0.95210307610577904                22                21
+   46                175.30685487399199               0.96342727836093855                25                24
+   47                207.58996658144758               0.96792078327336106                29                28
+   48                243.55454634579303               0.97222210755453431                34                33
+   49                281.02002406512781                 0.978086643140751                40                40
+   50                319.71391372262485                0.9830583563461136                49                50
+   51                415.17993316289517               0.98611105377726715                57                64
+   52                468.97997593487219               0.98904332157037556                66                76
+   53                  525.781814616828               0.99178465914661207                82                88
+   54                611.10744174509944               0.99401288451322234               107               121
+   55                669.64013367420966               0.99560906212398237               138               144
+   56                798.77258050403202               0.99675909020488263               160               171
+   57                 860.5721725547503               0.99842791078518778               188               254
+   58                937.95995186974437               0.99913197311136637               256               339
+   59                974.21818538317211               0.99951531212398237               394               514
+   60                999.46473080965086               0.99999581820945038               718              1025
